Reduce iteration limit for Newton's method solve in min curve/surf distance.
Newton's method exhibits quadratic convergence when started near the solution.
In practice, it seldom takes more than three or four iterations.  Each
iteration can add several significant figures of precision to the solution.

If it is started far from the solution, it will diverge or converge to a
maximum (or saddle point).  Both of these will happen quickly.

In short, there is never a reason to run Newton's method for hundreds
of iterations.