Fix broken tests uncovered by fixing Newton's method norm.
The sign of the u_contra and v_contra terms produced a point just above
the surface rather than beyond the surface.  Consequently, it was not
a proper geometric construction of a known nearest point off the surface.
Instead, the solution to the specified point was somewhat inside the
field.

In addition, the Newton's method norm fix makes a case that only failed
for float types work.  Consequently, we can remove the special case
handling of the floating point value.