Initial piecewise cubic spline skinning surface creator.
This started as a direct copy of the piecewise general surface creator,
with the general curve creator changed out for the PCHIP curve creator.

The rib data structure is overkill for this, but has been maintained
for now for similarity of interface.

It should be possible to support other cubic spline types without much
trouble.